tests/wpt/web-platform-tests/element-timing/observe-svg-image.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Element Timing: observe image inside SVG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
let beforeRender;
async_test(function (t) {
  if (!window.PerformanceElementTiming) {
    assert_unreached("PerformanceElementTiming is not implemented");
  }
  const observer = new PerformanceObserver(
    t.step_func_done(function(entryList) {
      assert_equals(entryList.getEntries().length, 1);
      const entry = entryList.getEntries()[0];
      const pathname = window.location.origin + '/element-timing/resources/circle.svg';
      checkElement(entry, pathname, 'my_svg', 'svg_id', beforeRender,
          document.getElementById('svg_id'));
      // Assume viewport has size at least 200, so the element is fully visible.
      checkRect(entry, [0, 200, 0, 200]);
      checkNaturalSize(entry, 200, 200);
    })
  );
  observer.observe({entryTypes: ['element']});
  beforeRender = performance.now();
}, "Able to observe svg image.");
</script>
<style>
body {
  margin: 0;
}
</style>
<svg width="300" height="300">
  <image href='resources/circle.svg' elementtiming='my_svg' id='svg_id'/>
</svg>
